backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "onchor-ai"
version = "0.1.0"
description = "Onchor.ai - Solidity Security Copilot with Persistent Collective Memory"
authors = [
    { name = "Onchor.ai Team" }
]
dependencies = [
    "click>=8.1.7",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.26.0",
    "web3>=6.15.1",
    "slither-analyzer>=0.10.0",
    "cognee>=1.0.0",
    "openai>=1.12.0",
    "instructor>=0.5.2",
    "tenacity>=8.2.3",
    "eth-account>=0.11.0",
    "anthropic>=0.17.0",
    "requests>=2.31.0"
]
requires-python = ">=3.10"

[project.scripts]
onchor-ai = "cli:cli"

[tool.setuptools]
py-modules = ["cli", "server"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*", "payments*", "memory*", "keeper*", "storage*", "onchor_ens*"]
